// src/app/metasystem/interfaces/app.profile.interface.ts
export interface AppProfile {
  SYSTEM_NAME: string;
  SYSTEM_SHORT_NAME: string;
  SYSTEM_SLOGAN: string;
  SYSTEM_PURPOSE: string;
  LOGO_PATH_LIGHT: string;
  LOGO_PATH_DARK: string;
  FAVICON_PATH: string;
  LOGO_WIDTH: string;
  LOGO_HEIGHT: string;
  LOGO_EXPAND_SIZE: string;
  LOGO_COLLAPSE_SIZE: string;
  LOGO_INLINE_GAP: string;
  LOGO_INLINE_MARGIN: string;
}